# optimize AtomicError