This commit optimizes AtomicError using atomic.Value. Benchmarks:

```
name                 old time/op  new time/op  delta
AtomicError/Load-6   305ns ±11%    12ns ± 6%  -96.18%  (p=0.000 n=10+10)
AtomicError/Set-6    314ns ±16%    14ns ± 2%  -95.61%  (p=0.000 n=10+9)
```